refactor(comment): 댓글 생성/삭제 원자성을 Processor로, 수정을 Writer로
CommentProcessor(@Transactional)가 createComment(댓글 insert + 게시글
댓글수 증가 + 부모 대댓글수 증가)와 deleteComment(soft/물리 삭제 + 부모
재귀 삭제 + 메타 카운트)를 한 tx로 원자화(C2). 재귀 삭제 헬퍼를 Service에서
이관. deleteComment는 postId(read-model)만 반환, Service가 응답 조립.

CommentWriter.update를 (memberId, commentId) 기반으로 바꿔 CommentReader를
주입, 자기 tx 안에서 readActiveOrThrow 재로드 후 작성자검증+updateContent로
detached 더티체킹 유실 방지(C1). CommentService의 클래스 @Transactional 제거.

미사용 increaseLikeCount/decreaseLikeCount 위임(reaction이 Writer 직접 호출)
제거. 오케스트레이션은 CommentProcessorTest, 수정 검증은 CommentWriterTest로.
